% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr}
\alias{clr}
\title{Class example CLR function}
\usage{
clr(x)
}
\arguments{
\item{x}{A vector of type numeric}
}
\value{
The centered log ratio transform of the input vector
}
\description{
Class example CLR function
}
