% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{Charlotte's log of summed exponents function}
\usage{
log_summed_exps(x)
}
\arguments{
\item{x}{A vector of type numeric.}
}
\value{
The log of summed exponents of the
input vector.
}
\description{
Charlotte's log of summed exponents function
}
\examples{
x <- 1:2000
log_summed_exps(x)



}
