% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{sample_size}
\alias{sample_size}
\title{sample size calculation: modification needed}
\usage{
sample_size(p0, p1, right_alpha = 0.025, power = 0.9)
}
\arguments{
\item{p0}{The proportion in H0}

\item{p1}{The proportion in H1}

\item{right_alpha}{The upper tail area}

\item{power}{Power of the test}
}
\value{
The sample size needed for a clinical trial that compares two proportions
}
\description{
sample size calculation: modification needed
}
