% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{Grant's implementation of log of sum of exponentiated values}
\usage{
log_summed_exps(x, base = exp(1))
}
\arguments{
\item{x}{a vector of type numeric}

\item{base}{a positive real number with which we
exponentiate. The default is Euler's number, e.}
}
\value{
the log of sum of exponentiated values
}
\description{
Grant's implementation of log of sum of exponentiated values
}
\examples{
log_summed_exps(1:2000)
log_summed_exps(c(1, 394, 102342, 18))

}
