% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{log summed exp}
\usage{
log_summed_exps(x)
}
\arguments{
\item{x}{vector of numeric values}
}
\value{
the value of log(sum(exp(x))
}
\description{
For example, let x=<1,2,3>. Then this function will return log(exp(1)+exp(2)+exp(3)). I think the HW is wrong, and you need to subtract the largest value
}
