% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{Taek's log-transformed summation of exponential terms function}
\usage{
log_summed_exps(x)
}
\arguments{
\item{x}{A vector of type numeric}
}
\value{
The log of summation of exponential values of the input vector
}
\description{
Taek's log-transformed summation of exponential terms function
}
\examples{
log_summed_exps(c(1:2000))
log_summed_exps(1)
log_summed_exps(rnorm(200))
}
