% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{Calculate the log-sum-exp of a numeric vector.}
\usage{
log_summed_exps(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the logarithm of the sum of exponentials of \code{x}
}
\description{
This function takes in a numeric vector and returns the log-sum-exp of its
elements, computed using a numerical trick to avoid overflow and underflow.
}
